Add comparator datastore.
This patch adds an in-memory datastore that is responsible for storing
the extracted files, and which source they came from, and what their
latest checksum is.  In addition, it also stores the edited files, the
files that were extracted from the current delta, and files that have
been deemed deleted.