Use target_path instead of file_name for output.
With deltas, because the initialized current_delta's file_name includes
a static number in front, the output statement for recreating the
cassette file suggests the incorrect file name, and could cause
confusion.  This change rectifies the problem by using `target_path`,
which is the absolute path to the file being regenerated.