test: rename {_/}spec (rspec) to {_/} test (minitest)
Many tools like `vim-test` infer testing frameworks based on the
filepattern.  Minitest is unique in the sense that it can take on
multiple forms (spec,test). It can be confusing to both contributors
and their tools to have files that look like rspec but that are actually
minitest's spec form using rspec's file pattern.